Exploratory Data Analysis (EDA) on the Titanic Dataset
Welcome to the repository dedicated to Exploratory Data Analysis (EDA) on the famous Titanic dataset. This resource is designed to help you understand the critical steps and techniques involved in EDA, a crucial stage in any data analysis project. The Titanic dataset serves as an excellent example for practicing EDA, as it contains valuable insights into the passengers on the ill-fated Titanic voyage.