Exploratory Data Analysis (EDA) on the Zomato Dataset
Welcome to the repository dedicated to Exploratory Data Analysis (EDA) on the Zomato dataset. This resource is designed to help you uncover insightful information about the restaurant industry, customer preferences, and trends in the world of dining. The Zomato dataset provides a rich source of data for conducting EDA and drawing meaningful conclusions.